finish testing and recs/insights button change
