+++ templates/view_music_recs.html
@@ -38,7 +38,7 @@
     <br>
     <div class="return-buttons">
         <a class="button" href="{{ url_for('home') }}">Return Home</a>
-        <a class="button" href="{{ url_for('music_recs') }}">Get New Insights</a>
+        <a class="button" href="{{ url_for('music_recs') }}">New Insights</a>
     </div>
         </div>
     </div>
