+++ templates/view_recommendation.html
@@ -37,7 +37,7 @@
 
     <div class="button-container">  
         <a class="button" href="{{ url_for('home') }}">Return Home</a>
-        <a class="button" href="{{ url_for('music_recs') }}">Get New Insights</a>
+        <a class="button" href="{{ url_for('music_recs') }}">New Insights</a>
     </div>
 </main>
